Implement AuthRepresentative
