+++ theoriq/api/v1alpha2/subscribe.py
@@ -1,17 +1,13 @@
 import logging
 import threading
 import time
-from typing import Callable, Optional
+from typing import Callable
 
 import httpx
-from biscuit_auth import PrivateKey
-from typing_extensions import Self
 
-from theoriq import AgentDeploymentConfiguration
 from theoriq.biscuit import AgentAddress
 
-from .protocol.biscuit_provider import BiscuitProvider, BiscuitProviderFromAPIKey, BiscuitProviderFromPrivateKey
-from .protocol.protocol_client import ProtocolClient
+from ..common import AuthRepresentative
 
 logger = logging.getLogger(__name__)
 
@@ -21,11 +17,7 @@ logger = logging.getLogger(__name__)
 SubscribeHandlerFn = Callable[[str], None]
 
 
-class Subscriber:
-    def __init__(self, biscuit_provider: BiscuitProvider, client: Optional[ProtocolClient] = None) -> None:
-        self._client = client or ProtocolClient.from_env()
-        self._biscuit_provider = biscuit_provider
-
+class Subscriber(AuthRepresentative):
     def new_job(
         self, agent_address: AgentAddress, handler: SubscribeHandlerFn, background: bool = False
     ) -> threading.Thread:
@@ -53,54 +45,3 @@ class Subscriber:
                 time.sleep(1)  # wait for 1 second before reconnecting
 
         return threading.Thread(target=_subscribe_job, daemon=background)
-
-    @classmethod
-    def from_api_key(cls, api_key: str, client: Optional[ProtocolClient] = None) -> Self:
-        """
-        Create a Subscriber from an API key.
-
-        Args:
-            api_key: The API key used for authentication
-            client: Optional protocol client, will create one from environment if not provided
-
-        Returns:
-            A new Subscriber instance configured with the API key credentials
-        """
-        protocol_client = client or ProtocolClient.from_env()
-        biscuit_provider = BiscuitProviderFromAPIKey(api_key=api_key, client=protocol_client)
-        return cls(biscuit_provider=biscuit_provider, client=protocol_client)
-
-    @classmethod
-    def from_agent(
-        cls, private_key: PrivateKey, address: Optional[AgentAddress] = None, client: Optional[ProtocolClient] = None
-    ) -> Self:
-        """
-        Create a Subscriber from an agent's private key and address.
-
-        Args:
-            private_key: The agent's private key used for authentication
-            address: Optional agent's address, will derive from a private key if not provided
-            client: Optional protocol client, will create one from environment if not provided
-
-        Returns:
-            A new Subscriber instance configured with the agent's credentials
-        """
-        protocol_client = client or ProtocolClient.from_env()
-        biscuit_provider = BiscuitProviderFromPrivateKey(
-            private_key=private_key, address=address, client=protocol_client
-        )
-        return cls(biscuit_provider=biscuit_provider, client=protocol_client)
-
-    @classmethod
-    def from_env(cls, env_prefix: str = "") -> Self:
-        """
-        Create a Subscriber from an agent's private key from environment variable.
-
-        Args:
-            env_prefix: Optional prefix for environment variable
-
-        Returns:
-            A new Subscriber instance configured with the agent's credentials
-        """
-        config = AgentDeploymentConfiguration.from_env(env_prefix=env_prefix)
-        return cls.from_agent(private_key=config.private_key)
